[BUD-126] review updates
