+++ cypress/integration/login.spec.ts
@@ -2,7 +2,7 @@ import { GraphQLError } from 'graphql';
 import schema from '../../server/src/schema.graphql';
 import { ROUTES } from '../../src/shared/routes';
 import { REQUEST_DELAY } from '../support/commands';
-import { UserRole } from '../../server/src/generated/schema-types';
+import { UserRole } from '../../src/types/schema-types';
 
 describe('Login Page', () => {
   beforeEach(() => {
