+++ src/__mocks__/context.ts
@@ -1,4 +1,4 @@
-import { AuthContextData } from 'context/AuthStore';
+import { AuthContextData } from 'contexts/AuthContext';
 
 export const authContext = (context?: Partial<AuthContextData>) => ({
   data: { role: 'BUDDY', token: 'dummy-token', userId: '1' },
