+++ src/components/AppRouter/__tests__/AppRouter.test.tsx
@@ -4,7 +4,7 @@ import { mockLocation, loginSuccessMock, authContext } from '__mocks__';
 import { MemoryRouter } from 'react-router-dom';
 import { MockedProvider } from '@apollo/react-testing';
 import { ROUTES } from 'shared/routes';
-import { AuthContext } from 'context/AuthStore';
+import AuthContext from 'contexts/AuthContext';
 import { UserRole } from 'types';
 import auth from 'utils/auth';
 import AppRouter from '../AppRouter';
